override the main function in gw_start instead of shadowing entire file
